RSDEV-1153 Extract PyRAT dialog from the TinyMCE plugin file
Move the React component out of Pyrat.tsx into PyratDialog.tsx so it can be
imported without the TinyMCE plugin registration side-effect; Pyrat.tsx now
holds only the plugin glue (mirrors the PubChem index.tsx / ImportDialog.tsx
split). Pure move: no behaviour change.
